requirements.txt
django
djangorestframework
google-api-python-client
oauth2client
paramiko
